requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
apturl==0.5.2
async-timeout==5.0.1
attrs==25.3.0
av==16.0.1
blinker==1.9.0
Brlapi==0.8.3
certifi==2025.11.12
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
ctranslate2==4.6.1
cupshelpers==1.0
cycler==0.12.1
datasets==4.0.0
dbus-python==1.2.18
defer==1.0.6
dill==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.3.0
fastapi==0.124.0
faster-whisper==1.2.1
filelock==3.18.0
Flask==3.0.0
flatbuffers==25.9.23
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.3.0
gunicorn==22.0.0
gyp==0.1
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httplib2==0.20.2
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib-metadata==4.6.4
iniconfig==2.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
keyboard==0.13.5
keyring==23.5.0
kiwisolver==1.4.9
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==3.0.2
matplotlib==3.10.5
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
netifaces==0.11.0
numpy==2.2.6
oauthlib==3.2.0
olefile==0.46
onnxruntime==1.23.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.6.0
propcache==0.3.2
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pyarrow==21.0.0
pycairo==1.20.1
pycparser==2.23
pycups==2.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
pytest==9.0.2
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.2.1
pytz==2025.2
pyxdg==0.27
PyYAML==6.0.2
regex==2025.7.34
reportlab==3.6.8
requests==2.32.4
safetensors==0.6.2
SecretStorage==3.3.1
selenium==4.38.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.3
sshuttle==1.0.5
starlette==0.50.0
sympy==1.14.0
systemd-python==234
tokenizers==0.21.4
tomli==2.3.0
tqdm==4.67.1
transformers==4.55.2
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
undetected-chromedriver==3.5.5
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
wadllib==1.3.6
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.3.2
xdg==5
xkit==0.0.0
xxhash==3.5.0
yarl==1.20.1
zipp==1.0.0
